QWEN2.5-7B-FEW_BUFFER_RATIO/requirement.txt
fire
vllm
lz4
numpy
matplotlib
pandas
pytz
seaborn
scipy
transformers
torch
huggingface_hub
tokenizers
safetensors
regex
tqdm
requests
urllib3
certifi
charset-normalizer
idna
filelock
fsspec
jinja2
markupsafe
pyyaml
accelerate
psutil
pillow
cycler
kiwisolver
pyparsing
python-dateutil
fonttools
contourpy
packaging
six
fire